% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlf.R
\name{inlf}
\alias{inlf}
\title{Implements \%in\% for lfactors}
\usage{
inlf(x, table)
}
\arguments{
\item{x}{same as \code{\link[base]{\%in\%}}}

\item{table}{same as \code{\link[base]{\%in\%}}}
}
\description{
Implements \%in\% for lfactors
}
\seealso{
\code{\link[base]{\%in\%}}
}

