% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfactors-package.R
\docType{package}
\name{lfactors-package}
\alias{lfactors-package}
\title{Allow a user to use == and \%in\% on an lfactor for both the label and the
level}
\description{
Similar to a factor, the user defines levels and labels when 
creating an \code{lfactor}. This is especially useful when
labels are long, and users know the levels well.
}
\details{
See \code{\link{lfactor}} for examples.
}
