% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llevels.R
\name{llevels}
\alias{llevels}
\title{Numeric Levels of an lfactor}
\usage{
llevels(x)
}
\arguments{
\item{x}{object of class lfactor}
}
\value{
A vector of levels
}
\description{
\code{llevels} gives the numeric levels of an lfactor.
}
\seealso{
\ifelse{latex}{\code{levels}}{\code{\link[base]{levels}}}
}
