\name{mlfactor}
\alias{mlfactor}
\alias{match,lfactor,ANY-method}
\alias{match,ANY,lfactor-method}
\alias{match,lfactor,lfactor-method}
\title{\code{\link[base]{match}} Function for lfactors}
\usage{
mlfactor(x, table, nomatch = NA_integer_, incomparables = NULL)
}
\arguments{
\item{x}{same as \ifelse{latex}{\code{match}}{\code{\link[base]{match}}}}

\item{table}{same as \ifelse{latex}{\code{match}}{\code{\link[base]{match}}}}

\item{nomatch}{same as \ifelse{latex}{\code{match}}{\code{\link[base]{match}}}}

\item{incomparables}{same as \ifelse{latex}{\code{match}}{\code{\link[base]{match}}}}
}
\description{
\ifelse{latex}{\code{match}}{\code{\link[base]{match}}} function for lfactors.
}
\details{
Allows \code{match} to work when the \code{x} or \code{table} arguments in a call to match are lfactors.
}
\seealso{
\ifelse{latex}{\code{match}}{\code{\link[base]{match}}}
}

