\name{inlf}
\alias{inlf}
\alias{\%in\%,ANY,lfactor-method}
\alias{\%in\%,lfactor,ANY-method}
\alias{\%in\%,lfactor,lfactor-method}
\title{Implements \%in\% for lfactors}
\usage{
inlf(x, table)
}
\arguments{
\item{x}{same as \ifelse{latex}{\code{\%in\%}}{\code{\link[base:match]{\%in\%}}}}

\item{table}{same as \ifelse{latex}{\code{\%in\%}}{\code{\link[base:match]{\%in\%}}}}
}
\description{
Implements \%in\% for lfactors.
}
\seealso{
\ifelse{latex}{\code{\%in\%}}{\code{\link[base:match]{\%in\%}}}
}

