\name{inlf}
\alias{inlf}
\alias{\%in\%,ANY,lfactor-method}
\alias{\%in\%,lfactor,ANY-method}
\alias{\%in\%,lfactor,lfactor-method}
\title{Implements \%in\% for lfactors}
\usage{
inlf(x, table)
}
\arguments{
\item{x}{same as \code{\link[base]{\%in\%}}}

\item{table}{same as \code{\link[base]{\%in\%}}}
}
\description{
Implements \%in\% for lfactors
}
\seealso{
\code{\link[base]{\%in\%}}
}

