\name{mlfactor}
\alias{mlfactor}
\alias{match,lfactor,ANY-method}
\alias{match,ANY,lfactor-method}
\alias{match,lfactor,lfactor-method}
\title{\code{\link[base]{match}} function for lfactors}
\usage{
mlfactor(x, table, nomatch = NA_integer_, incomparables = NULL)
}
\arguments{
\item{x}{same as \code{\link[base]{match}}}

\item{table}{same as \code{\link[base]{match}}}

\item{nomatch}{same as \code{\link[base]{match}}}

\item{incomparables}{same as \code{\link[base]{match}}}
}
\description{
\code{\link[base]{match}} function for lfactors
}
\details{
Allows match to work when the `x' or `table' arguments in a call to match are lfactors.
}
\seealso{
\code{\link[base]{match}}
}

