% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlfactor.R
\name{mlfactor}
\alias{mlfactor}
\title{\code{\link[base]{match}} function for lfactors}
\usage{
mlfactor(x, table, nomatch = NA_integer_, incomparables = NULL)
}
\arguments{
\item{x}{same as \code{\link[base]{match}}}

\item{table}{same as \code{\link[base]{match}}}

\item{nomatch}{same as \code{\link[base]{match}}}

\item{incomparables}{same as \code{\link[base]{match}}}
}
\description{
\code{\link[base]{match}} function for lfactors
}
\details{
Allows match to work when the `x' or `table' arguments in a call to match are lfactors.
}
\seealso{
\code{\link[base]{match}}
}

